minio_helper/requirements.txt
boto3>=1.28,<2
PyYAML>=6,<7
